wxauto_mgt/web/requirements.txt
fastapi>=0.95.0
uvicorn>=0.22.0
pydantic>=1.10.7
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
aiofiles>=23.1.0
